QHDEV-734 - Move collapsible and popover into src/behaviours
`collapsible.js` and `popover/` are interaction engines shared by several
components: collapsible drives the accordion, the breadcrumb overflow
menu and the mega main navigation; popover drives the tool tip and the
toggle tip. Neither is owned by a component, but both lived under
`src/components/_global/js/`, which made `_global` read as a component
that happened to export machinery for others.

They now sit in `src/behaviours/`, a layer of their own between
`src/utils` and `src/components`. A README records the distinction that
decides where a new module goes: a util knows nothing about our markup;
a behaviour knows QLD conventions (class names, `dataLayer` events,
`data-` attribute contracts) but is not owned by one component; anything
used by a single component belongs in that component.

`collapsible` drops its private `toItems` in favour of the identical
`dom.toElementArray`, which is the same normalisation the animate engine
already needed.

Otherwise the module bodies are unchanged.
